Extract base classes for various ODataQueryResult(s)
