+++ src/Microsoft.Restier.WebApi/Results/EntityCollectionResult.cs
@@ -11,7 +11,7 @@ namespace Microsoft.Restier.WebApi.Results
     /// <summary>
     /// Represents a collection of entity instances being returned from an action.
     /// </summary>
-    public class EntityCollectionResult : ODataQueryResult
+    public class EntityCollectionResult : CollectionQueryResult
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="EntityCollectionResult" /> class.
@@ -20,23 +20,8 @@ namespace Microsoft.Restier.WebApi.Results
         /// <param name="edmType">The EDM type reference of the entities.</param>
         /// <param name="context">The context where the action is executed.</param>
         public EntityCollectionResult(IQueryable query, IEdmTypeReference edmType, DomainContext context)
-            : base(edmType)
+            : base(query, edmType, context)
         {
-            Ensure.NotNull(query, "query");
-            Ensure.NotNull(context, "context");
-
-            this.Query = query;
-            this.Context = context;
         }
-
-        /// <summary>
-        /// Gets the query that returns a collection of entities.
-        /// </summary>
-        public IQueryable Query { get; private set; }
-
-        /// <summary>
-        /// Gets the context where the action is executed.
-        /// </summary>
-        public DomainContext Context { get; private set; }
     }
 }
