+++ src/Microsoft.Restier.WebApi/Results/EntityResult.cs
@@ -12,7 +12,7 @@ namespace Microsoft.Restier.WebApi.Results
     /// <summary>
     /// Represents a single entity instance being returned from an action.
     /// </summary>
-    public class EntityResult : ODataQueryResult
+    public class EntityResult : SingleQueryResult
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="EntityResult" /> class.
@@ -21,24 +21,8 @@ namespace Microsoft.Restier.WebApi.Results
         /// <param name="edmType">The EDM type reference of the entity.</param>
         /// <param name="context">The context where the action is executed.</param>
         public EntityResult(IQueryable query, IEdmTypeReference edmType, DomainContext context)
-            : base(edmType)
+            : base(query, edmType, context)
         {
-            Ensure.NotNull(query, "query");
-            Ensure.NotNull(context, "context");
-
-            this.Context = context;
-
-            this.Result = query.SingleOrDefault();
         }
-
-        /// <summary>
-        /// Gets the result object.
-        /// </summary>
-        public object Result { get; private set; }
-
-        /// <summary>
-        /// Gets the context where the action is executed.
-        /// </summary>
-        public DomainContext Context { get; private set; }
     }
 }
